Fixed up the namespace and moved some code around
