+++ VisualStudioSolutionGenerator.Tests/MSBuildUtilitiesTests.cs
@@ -12,8 +12,6 @@ namespace VisualStudioSolutionGenerator.Tests
 
     using NUnit.Framework;
 
-    using SolutionGenerator;
-
     [TestFixture]
     [Parallelizable(ParallelScope.All)]
     public class MSBuildUtilitiesTests
