+++ VisualStudioSolutionGenerator/VisualStudioSolutionGenerator.csproj
@@ -24,6 +24,7 @@
     <Compile Include="Options.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="SolutionGenerator.cs" />
     <Compile Include="SolutionUtilities.cs" />
   </ItemGroup>
   <ItemGroup>
